more theming fixes
