+++ packages/experimental/framework/src/themeHelper.ts
@@ -1,12 +1,9 @@
-import * as React from 'react';
 import { ThemeHelper } from '@fluentui-react-native/use-styling';
-import { Theme, ThemeContext } from '@fluentui-react-native/theme-types';
+import { Theme, useTheme } from '@fluentui-react-native/theme-types';
 import { defaultFluentTheme } from '@fluentui-react-native/default-theme';
 
 export const themeHelper: ThemeHelper<Theme> = {
-  useTheme: () => {
-    return React.useContext(ThemeContext) || defaultFluentTheme;
-  },
+  useTheme: () => useTheme() || defaultFluentTheme,
   getComponentInfo: (theme: Theme, name: string) => {
     const components = theme.components || {};
     return components[name];
